+++ privval/socket.go
