+++ privval/socket_tcp.go
