+++ privval/socket_test.go
